refactor: Moved `state` folder out of `colorizer`
